Create `NumericalFormatter` (#1007)
* Add NumericalFormatter + tests

* Integrate NumericalFormatter into DataProcessor

* Update `representation` values to match the ones from RDT's `computer_representation`

* Add tests + update logic

* Typo fix

* Fix test mock bug

* Fix minor column

* Fix rebase + format_data takes a Series

* Remove test added by accident

+++ tests/unit/metadata/test_metadata_upgrader.py
@@ -663,7 +663,7 @@ def test_convert_metadata():
             },
             'salary': {
                 'sdtype': 'numerical',
-                'representation': 'int64'
+                'representation': 'Int64'
             },
             'duration': {
                 'sdtype': 'categorical'
@@ -673,7 +673,7 @@ def test_convert_metadata():
             },
             'high_perc': {
                 'sdtype': 'numerical',
-                'representation': 'float64'
+                'representation': 'Float'
             },
             'placed': {
                 'sdtype': 'boolean'
@@ -752,7 +752,7 @@ def test_convert_metadata_with_constraints(upgrade_constraints_mock):
         'columns': {
             'salary': {
                 'sdtype': 'numerical',
-                'representation': 'int64'
+                'representation': 'Int64'
             },
             'student_id': {
                 'sdtype': 'numerical'
